+++ packages/core/tsup.config.ts
@@ -8,6 +8,13 @@ export default defineConfig([
     external: ['react', 'svelte', 'vue'],
     dts: true
   },
+  {
+    entry: ['prompts/index.ts'],
+    format: ['cjs', 'esm'],
+    external: ['react', 'svelte', 'vue'],
+    outDir: 'prompts/dist',
+    dts: true
+  },
   // React APIs
   {
     entry: ['react/index.ts'],
@@ -16,14 +23,14 @@ export default defineConfig([
       js: "'use client'"
     },
     format: ['cjs', 'esm'],
-    external: ['react', 'svelte', 'vue'],
+    external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true
   },
   {
     entry: ['react/index.server.ts'],
     outDir: 'react/dist',
     format: ['cjs', 'esm'],
-    external: ['react', 'svelte', 'vue'],
+    external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true
   },
   // Svelte APIs
@@ -32,7 +39,7 @@ export default defineConfig([
     outDir: 'svelte/dist',
     banner: {},
     format: ['cjs', 'esm'],
-    external: ['react', 'svelte', 'vue'],
+    external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
     // `sswr` has some issue with `.es.js` that can't be resolved correctly by
     // vite so we have to bundle it here.
@@ -44,7 +51,16 @@ export default defineConfig([
     outDir: 'vue/dist',
     banner: {},
     format: ['cjs', 'esm'],
-    external: ['react', 'svelte', 'vue'],
+    external: ['react', 'svelte', 'vue', 'solid-js'],
+    dts: true
+  },
+  // Solid APIs
+  {
+    entry: ['solid/index.ts'],
+    outDir: 'solid/dist',
+    banner: {},
+    format: ['cjs', 'esm'],
+    external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true
   }
 ])